+++ src/WheelEncoder.h
@@ -22,7 +22,7 @@ class WheelEncoder : public WheelBase {
 		void attachEncoder(EncoderBase * encoder);						//Attach the Encoder
 		void attachPid(Pid * pid);									//Attach the Pid
 		
-    	virtual void move(float velocity);                          //velocity demanded radians per second.
+    	virtual void move(double velocity);                          //velocity demanded radians per second.
     	virtual void stop();										//reset duty to 0 and direction to forward
 		virtual void update();
 
@@ -33,7 +33,7 @@ class WheelEncoder : public WheelBase {
 		EncoderBase * encoder_;											//Encoder	
 		Pid * pid_;													//Pid
 		os_timer_t timer_;											//Timer for periodic task, Pid controller	
-		float period_pid_controller_;								//Period of the periodic task, Pid controller
+		double period_pid_controller_;								//Period of the periodic task, Pid controller
 		static void timer_callback_(void *pArg);					//Timer callback function, Pid controller
 		void timer_setup_();										//Setup the timer for the periodic task, Pid controller
 		void update_();												//Pid controller function
